# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample API documentation using Swagger in Go.",
        "title": "Event-Trigger",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "http://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/": {
            "get": {
                "description": "Health check endpoint",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "Health"
                ],
                "summary": "Ping the server",
                "responses": {
                    "200": {
                        "description": "Event Trigger Platform running.",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/trigger": {
            "put": {
                "description": "Creates a trigger that calls an API at a scheduled time",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "Triggers"
                ],
                "summary": "Edit a trigger",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Enter id to edit trigger",
                        "name": "triggerId",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Trigger Data",
                        "name": "trigger",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.EditTrigger"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.EditTrigger"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a trigger that calls an API at a scheduled time",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Triggers"
                ],
                "summary": "Create a new trigger",
                "parameters": [
                    {
                        "description": "Trigger Data",
                        "name": "trigger",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.Trigger"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.Trigger"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Creates a trigger that calls an API at a scheduled time",
                "consumes": [
                    "text/plain"
                ],
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "Triggers"
                ],
                "summary": "Delete a trigger by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Enter id for trigger deletion",
                        "name": "triggerId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/trigger-test": {
            "post": {
                "description": "Test event trigger",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Trigger-Test"
                ],
                "summary": "Test trigger",
                "parameters": [
                    {
                        "description": "Trigger Data",
                        "name": "trigger",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.Trigger"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.Trigger"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/triggers": {
            "get": {
                "description": "Creates a trigger that calls an API at a scheduled time",
                "consumes": [
                    "text/plain"
                ],
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "Triggers"
                ],
                "summary": "Get all triggers for a user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Enter userId",
                        "name": "userId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handler.Trigger"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.EditTrigger": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "phoneNo": {
                    "type": "string"
                },
                "schedule": {
                    "type": "string"
                }
            }
        },
        "handler.Trigger": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "phoneNo": {
                    "type": "string"
                },
                "schedule": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        }
    }
}